Changed Ray Amount variables  types to int from short
C# was doing an incorrect boxing of the value type when being boxed to a certain int variable.
This was causing some values that should of been true for IsColliding in a couple different directions to be false instead of true.

This was because the int value was being boxed just straight to 0 making know rays to be sent out at all.

+++ Runtime/Physics/Collision/CollisionController.cs
@@ -10,8 +10,8 @@ namespace SF.Physics.Collision
 		public float VerticalRayDistance;
 
 		[Header("Ray Amount")]
-		public short HoriztonalRayAmount;
-		public short VerticalRayAmount;
+		public int HoriztonalRayAmount;
+		public int VerticalRayAmount;
 
 		[HideInInspector] public RaycastHit2D[] RaycastHit2Ds;
 		public CollisionController(float horiztonalRayDistance = 0.01f,
